rename utils and test_utils to inline_markdown and test_inline_markdown
